% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{advancedBayesianCausalInference}
\alias{advancedBayesianCausalInference}
\title{Advanced Bayesian Causal Inference using Rcpp}
\usage{
advancedBayesianCausalInference(data, iterations, burnin)
}
\arguments{
\item{data}{A data frame or matrix containing the data for analysis.}

\item{iterations}{An integer specifying the number of iterations for the MCMC algorithm.}

\item{burnin}{An integer specifying the number of burn-in iterations to discard.}
}
\value{
A list containing the results of the Bayesian causal inference analysis.
}
\description{
This function performs advanced Bayesian causal inference using Rcpp for efficient computation.
}
