% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausalRF.R
\name{causal_random_forest}
\alias{causal_random_forest}
\title{Causal Random Forest}
\usage{
causal_random_forest(
  X,
  Y,
  W,
  num.trees = 500,
  mtry = NULL,
  min.node.size = 5,
  sample.fraction = 0.632,
  alpha = 0.05
)
}
\arguments{
\item{X}{A data frame or matrix of covariates (features).}

\item{Y}{A numeric vector of outcomes (response variable).}

\item{W}{A binary vector indicating treatment assignment (1 for treated, 0 for control).}

\item{num.trees}{The number of trees to grow in the random forest.}

\item{mtry}{The number of features to consider at each split.}

\item{min.node.size}{The minimum size of terminal nodes.}

\item{sample.fraction}{The fraction of samples to use for each tree.}

\item{alpha}{The significance level for the confidence intervals.}
}
\value{
A list containing estimated treatment effects and other relevant statistics.
}
\description{
This function implements a Causal Random Forest algorithm to estimate treatment effects from observational data. It allows for the estimation of heterogeneous treatment effects by leveraging the random forest framework.
}
