\name{SA23204151-package}
\alias{SA23204151-package}
\alias{SA23204151}
\docType{package}
\title{This package provides several functions for causal identification}
\description{\packageDescription{SA23204151}}
\section{Package Content}{\packageIndices{SA23204151}}
\author{
Rui Wu
Maintainer: Rui Wu <wurui22@mail.ustc.edu.cn>
}
\keyword{package}
